+++ src/main/java/org/ahpuh/surf/follow/domain/FollowRepository.java
@@ -1,6 +1,5 @@
-package org.ahpuh.surf.follow.repository;
+package org.ahpuh.surf.follow.domain;
 
-import org.ahpuh.surf.follow.entity.Follow;
 import org.ahpuh.surf.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
