+++ src/main/java/org/ahpuh/surf/user/domain/User.java
@@ -4,7 +4,7 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.ahpuh.surf.category.domain.Category;
 import org.ahpuh.surf.common.entity.BaseEntity;
-import org.ahpuh.surf.follow.entity.Follow;
+import org.ahpuh.surf.follow.domain.Follow;
 import org.ahpuh.surf.like.entity.Like;
 import org.ahpuh.surf.post.domain.Post;
 import org.ahpuh.surf.user.dto.request.UserUpdateRequestDto;
