+++ src/test/java/org/ahpuh/surf/integration/follow/controller/FollowControllerTest.java
@@ -1,8 +1,8 @@
 package org.ahpuh.surf.integration.follow.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.ahpuh.surf.follow.entity.Follow;
-import org.ahpuh.surf.follow.repository.FollowRepository;
+import org.ahpuh.surf.follow.domain.Follow;
+import org.ahpuh.surf.follow.domain.FollowRepository;
 import org.ahpuh.surf.user.controller.UserController;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserRepository;
