+++ src/test/java/org/ahpuh/surf/integration/post/repository/PostQueryDslTest.java
@@ -3,8 +3,8 @@ package org.ahpuh.surf.integration.post.repository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.ahpuh.surf.category.domain.Category;
 import org.ahpuh.surf.category.domain.CategoryRepository;
-import org.ahpuh.surf.follow.entity.Follow;
-import org.ahpuh.surf.follow.repository.FollowRepository;
+import org.ahpuh.surf.follow.domain.Follow;
+import org.ahpuh.surf.follow.domain.FollowRepository;
 import org.ahpuh.surf.post.domain.Post;
 import org.ahpuh.surf.post.domain.repository.PostRepository;
 import org.ahpuh.surf.post.dto.ExploreDto;
